refactore server api
